Recommendations: moves component into recommendations/ folder and updates imports
