+++ src/components/about/recommendations/Recommendations.jsx
@@ -1,6 +1,5 @@
 import React from 'react'
-
-import { WebLink } from '../common/WebLink'
+import { WebLink } from '../../common/WebLink'
 
 export const Recommendations = () => {
   return (
